Freshly Graduated and Looking for Collaborators
I've just graduated as a Software Engineer from the Tech Talent Academy coding Bootcamp 🥇 . What an experience and thanks to V for her patience and for always being there for us. Also thanks to my project partners @georgjnrr and @naqashb7 and everyone else in the group.

So now the real journey begins. 

My pet project is to build an LMS for change makers and change practitioners. Come join me.